fix: #20 수업일지 작성 기능 이관
