+++ src/routes/classes.api.ts
@@ -15,9 +15,11 @@ router.put("/:classId", authMiddleware, classesController.updateClass);
 router.delete("/:classId", authMiddleware, classesController.deleteClass);
 // 출석 조회
 router.get("/:classId/attendance", authMiddleware, classesController.getAttendance);
-// 출석 기록
+// 출석 체크
 router.post("/:classId/attendance", authMiddleware, classesController.checkAttendance);
 // 수업 보고서 조회
 router.get("/:classId/report", authMiddleware, classesController.getClassReport);
+// 수업 보고서 생성
+router.post("/:classId/report", authMiddleware, classesController.createClassReport);
 
 export default router;
