--- src/types/classReports.types.ts
@@ -1,5 +1,0 @@
-import { Context } from "hono";
-
-export interface ClassReportsController {
-  createClassReport: (c: Context) => Promise<Response>;
-}
